Removed Game and commandWords class, and moved the 'move' method out of board class
For milestone 2 the loop and parser in the game class is unnecessary as it will be taken care of in the view of the MVC so it was deleted. The TA said the board class should delgate to the rabbit fox classes to move themselves and so that was implemented. The TestModel class is just to test the model until the VIew and controller work