Got save and load working
Added GameSaver, GameLoader, and SaveFile classes. Board is first converted to a string 2D array representation of the board before being saved as an XML file. XML is converted back to a string array and then a new board is created based on that string 2D array. SaveFile is just a POJO for serialization purposes. addGamePieces() function in board was changed to protected so it can be accessed by GameLoader. Save and load functions should be integrated into the rest of the game already. View and Controller classes slightly changed to accommodate for integration of save and load features and future integration of level builder.