Almost done. Potential oversight in either Board or Controller classes.
Everything is working except for start button when successfully creating a board in level builder. When starting the game straight from a successful level build, there will be a null pointer exception. I suspect it has something to do with either the controller or something with one of the constructors in the board.